Modifiche generali
Tema scuro attivabile
Inserimento di misurazioni standardizzate per bottoni classici e bottoni link
Nuova metodologia di inserimento delle icone
Pagine responsive

+++ Database/db-rifugio.sql
@@ -36,7 +36,7 @@ CREATE TABLE AnimaleRifugio (
 	Storia TEXT NOT NULL,
 	Sesso ENUM ('M', 'F') NOT NULL,
 	Peso FLOAT(2, 1) NOT NULL,
-	Eta FLOAT(2, 1) NOT NULL,
+	Eta FLOAT(2, 2) NOT NULL,
 	PthImg TEXT NOT NULL UNIQUE,
 	Colore VARCHAR(10) NOT NULL,
 	Caratteristiche JSON NOT NULL,
